+++ core/changelog.md
@@ -1,2 +1,3 @@
 - fix: allow setting authorization header through extra headers and in allow list and deny list
 - feat: added support for gemini google search tool
+- fix: function response part handling in gemini